+++ changelog.md
@@ -1,3 +1,7 @@
 # bpd toolkit
+# [0.1.4] WIP
+# [0.1.3] 2020-10-11
+* [Added] Debounce function and class
+* [Added] Throttle function and class
 # [0.1.1] 2020-09-23
 * Initial release, added functions and test cases